vincent 1.0 typescript work salvaged from extension installation
this is a confusing directory structure and probably because I salvaged it from
the extension installation directory rather from the original repo, which was lost
It is here for historical purposes and maybe some of the typescript
conversion work can be resurrected sometime. Meanwhile I'm branching back to the
pure javascript version which is understandable to me.